crypto-ai-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-wtf==1.2.1
wtforms==3.1.1
flask-limiter==3.5.0
marshmallow==3.20.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
alembic==1.12.1
flask-jwt-extended==4.5.3
bcrypt==4.1.1
flask-socketio==5.3.6
pandas==2.1.4
numpy==1.26.2
ccxt==4.1.22
ta==0.11.0
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4
matplotlib==3.8.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
setuptools==69.0.2
werkzeug==3.0.1
cryptography==41.0.7